Added gem simple_form
Added simple form to build my existing forms with bootstrap. This is a
very quick and effective way!

Should be used as an alternative to using LESS for generating bootstrap
themes on existing :resources. In Windows environment its not easy to
install 'therubyracer' gem or even 'libv8' for that matter! Thus it
becomes hard for Windows users to generate bootstrap themed
[:resources].

Using simple_form is a great, easy and lightweight solution for Windows
users.